Refactor build tools
With this refactor we split the SDK code from the example_algo
code. The script build_rt_images.sh takes as argument the root
directory of the algorithm project. That root directly is expected
to have a build.sh file

Currently the algorithm project has to take care of building the
final runtime image. The example project is provided here as guidance
for algorithm authors. We will move to a different model, in which
algorithms don't have to provide Dockerfile in the future